pyproject.toml
[project]
name = "reve"
version = "0.1.0"
description = "Codebase for REVE: A Foundation Model for EEG"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.5.2",
    "h5py>=3.13.0",
    "hydra-core>=1.3.2",
    "idr-torch>=2.4.0",
    "linear-attention-transformer>=0.19.1",
    "lmdb>=1.6.2",
    "mne>=1.9.0",
    "moabb>=1.4.3",
    "numpy>=2.2.4",
    "openpyxl>=3.1.5",
    "optuna>=4.7.0",
    "pandas>=2.2.3",
    "peft>=0.15.1",
    "pyedflib>=0.1.40",
    "rich>=13.9.4",
    "safetensors>=0.5.3",
    "scikit-learn>=1.6.1",
    "scipy>=1.15.2",
    "setuptools>=78.1.0",
    "timm>=1.0.19",
    "torch>=2.6.0",
    "transformers==4.56.2",
    "wandb>=0.19.8",
]

[project.optional-dependencies]
flash-attn = ["flash-attn"]


[tool.ruff]

line-length = 120
indent-width = 4

[tool.ruff.lint]
extend-select = [
    "ARG", # flake8-unused-arguments
    "C90", # mccabe complexity
    "COM", # commas 
    "C4",  # flake8-comprehensions
    "E",   # pycodestyle
    "F",   # Pyflakes
    "I",   # isort
    "PL",  # pylint
    "SIM", # flake8-simplify
    "TID", # flake8-tidy-imports
    "W",   # pycodestyle warnings
]
ignore = ["PLR2004"]
fixable = ["ALL"]

# Ignore import violations in all `__init__.py` files.
[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402", "F401", "F403", "F811"]
"old_ref/**" = ["ALL"]
"src/utils/stable_adamw_distributed.py" = ["ALL"]
"src/utils/stable_adamw.py" = ["ALL"]
"hf/**" = ["ALL"]

[tool.ruff.lint.isort]
lines-after-imports = 2

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.pyrefly]
project-excludes = [
    # external code
    "preprocessing",
    "src/utils/stable_adamw.py",
    "src/utils/stable_adamw_distributed.py",
    # legacy code
    "old_ref",
    # exported code
    "hf",
]

[tool.pyrefly.errors]
untyped-import = false
